Refactor blog views and card components for better modularity
Centralized blog view logic into a reusable `ViewBlogComponent`. Modified `CardComponent` to include navigation links and improved visual structure. Adjusted type definitions to align with changes and introduced the `useBodyClass` hook for managing page-specific body styles.
